refactor: module structure
I restructured the code base to better represent how I'm envisioning
this to work.

In essence, I pulled out lexing and parsing into a dedicated module
(lexparse). This is because eventually I want to be able to do
all the computations up to that stage for individual files, then merge
the output of each file into a single journal. This is basically the
idea behind #12. Also, I pulled out the compile stage into the root
module.

closes: #21
