+++ src/contracts/output.rs
