+++ src/lexparse/contracts.rs
@@ -1,2 +1,1 @@
-pub mod ast;
 pub mod tokens;
