+++ src/lexparse/lex/basic.rs
@@ -1,4 +1,4 @@
-use crate::input::contracts::tokens::{Keyword, Token};
+use crate::lexparse::contracts::tokens::{Keyword, Token};
 use nom::branch::alt;
 use nom::bytes::complete::{tag, take_until};
 use nom::character::complete::none_of;
@@ -6,7 +6,7 @@ use nom::combinator::{opt, peek};
 use nom::sequence::{delimited, preceded};
 use nom::Parser;
 
-use crate::input::lex::{core::LexResult, whitespace};
+use crate::lexparse::lex::{core::LexResult, whitespace};
 
 pub fn lex_indent(input: &str) -> LexResult<'_, Token> {
     let (input, _) = tag("  ").parse(input)?;
