+++ src/lexparse/lex/timestamp.rs
@@ -1,5 +1,5 @@
-use crate::input::contracts::tokens::{Timestamp, Token};
-use crate::input::lex::core::LexResult;
+use crate::lexparse::contracts::tokens::{Timestamp, Token};
+use crate::lexparse::lex::core::LexResult;
 
 use chrono::{DateTime, FixedOffset, NaiveDate, NaiveDateTime, NaiveTime};
 use nom::bytes::complete::take;
