+++ src/lib.rs
@@ -1,3 +1,4 @@
-pub mod input;
-pub mod output;
+pub mod compile;
+pub mod contracts;
+pub mod lexparse;
 pub mod utils;
