+++ tests/test_successful.rs
@@ -1,8 +1,8 @@
 use boki;
 
-fn compile(fp: &str) -> boki::input::InputResult<boki::output::Journal> {
+fn compile(fp: &str) -> boki::lexparse::InputResult<boki::contracts::output::Journal> {
     let content = std::fs::read_to_string(fp).expect("Could not open file.");
-    boki::input::compile_string(&content)
+    boki::lexparse::compile_string(&content)
 }
 
 #[rstest::rstest]
@@ -13,7 +13,7 @@ fn test_successful(#[case] case: &str) {
     let result = compile(&format!("tests/cases/{case}.boki")).expect("Failed.");
     let expected_str = std::fs::read_to_string(&format!("tests/cases/{case}.output.json"))
         .expect("Could not read output file.");
-    let expected: boki::output::Journal =
+    let expected: boki::contracts::output::Journal =
         serde_json::from_str(&expected_str).expect("Could not parse output file.");
     assert_eq!(result, expected);
 }
